Buffers now use VMA for allocations

+++ src/Vulkan/VulkanPipeline.cpp
@@ -635,7 +635,7 @@ VulkanPipeline::VulkanPipeline(std::shared_ptr<VulkanDevice> device, VkFormat co
             .sType = VK_STRUCTURE_TYPE_PIPELINE_LAYOUT_CREATE_INFO,
             .setLayoutCount = static_cast<uint32_t>(m_DescriptorSetLayouts.size()),
             .pSetLayouts = m_DescriptorSetLayouts.data(),
-            .pushConstantRangeCount = pushConstantRanges.size(),
+            .pushConstantRangeCount = (uint32_t) pushConstantRanges.size(),
             .pPushConstantRanges = pushConstantRanges.data(),
     };
